requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
fonttools==4.51.0
idna==3.7
importlib_resources==6.4.0
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
numpy==1.26.0
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pathlib==1.0.1
pwinput==1.0.3
pydicom==2.4.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyxnat==1.6.2
requests==2.32.2
setuptools==70.0.0
six==1.16.0
tabulate==0.9.0
tzdata==2024.1
urllib3==2.2.1
wheel==0.43.0
zipp==3.18.2
openpyxl
